Enhance GoogleProvider and MediaGenerationHandler error handling
- Added checks in GoogleProvider to handle responses indicating blocked content, throwing appropriate exceptions.
- Refactored MediaGenerationHandler to utilize MediaErrorMessageBuilder for generating user-friendly error messages.
- Updated tests for GoogleProvider to validate blocked content handling and ensure proper exception throwing.
