+++ PR-610-Review.md
@@ -1,0 +1,96 @@
+# PR #610 Review – Enhance error handling for blocked content
+
+**Branch:** `fix/image-errors` → `main`
+**Dateien:** 4 geändert (+374 / -26)
+
+---
+
+## 1. `chat()` und `chatStream()` prüfen `finishReason` nicht
+
+`checkGeminiFinishReason()` wird nur in den Image-Methoden aufgerufen (`generateImageWithGemini`, `generateImageFromImagesWithGemini`). Die Methoden `chat()` und `chatStream()` lesen die Response aber ohne jegliche Prüfung:
+
+```php
+// chat() – Zeile 145
+return $data['candidates'][0]['content']['parts'][0]['text'] ?? '';
+```
+
+Wenn Gemini hier `finishReason: SAFETY` zurückgibt, wird einfach ein leerer String zurückgegeben – kein Fehler, kein Logging, kein Hinweis an den User. Das ist inkonsistent zum neuen Pattern und potenziell ein stiller Bug.
+
+**Empfehlung:** `checkGeminiFinishReason()` auch in `chat()` und `chatStream()` aufrufen.
+
+---
+
+## 2. `generateVideo()` nutzt das neue Pattern nicht
+
+In `generateVideo()` (Zeile 687) wird Content-Blocking per String-Matching erkannt:
+
+```php
+if (str_contains(strtolower($errorMessage), 'safety') || str_contains(strtolower($errorMessage), 'blocked')) {
+    throw new \Exception('Video generation blocked by safety filters: '.$errorMessage);
+}
+```
+
+Es wird eine generische `\Exception` geworfen statt `ProviderException::contentBlocked()`. Dadurch greift die neue `buildErrorMessage()`-Logik im `MediaGenerationHandler` nicht, und der User bekommt nur die generische Fehlermeldung.
+
+**Empfehlung:** Auch hier `ProviderException::contentBlocked()` verwenden.
+
+---
+
+## 3. Hardcoded Translations im Backend statt i18n
+
+`getBlockReasonExplanations()` enthält ~45 Zeilen hardcoded Übersetzungen (DE/EN) direkt im PHP-Code. Das widerspricht dem Projekt-Pattern, wo User-facing Text über `vue-i18n` (`en.json` / `de.json`) verwaltet wird.
+
+Probleme:
+- Translations sind über Backend-Code verstreut statt zentral in den JSON-Dateien
+- Neue Sprachen erfordern Code-Änderungen statt JSON-Einträge
+- Übersetzer haben keinen Zugriff auf PHP-Dateien
+
+**Empfehlung:** Entweder die Messages als Übersetzungs-Keys zurückgeben und im Frontend übersetzen, oder – wenn Backend-Translation unvermeidbar ist – in eine eigene Translation-Service-Klasse auslagern.
+
+---
+
+## 4. MediaGenerationHandler ist zu groß (868 Zeilen)
+
+Der Handler hat jetzt 868 Zeilen. Das PR fügt weitere ~120 Zeilen hinzu, davon ~90 Zeilen reine Übersetzungstexte. Laut Projekt-Conventions sollten Services unter 500 Zeilen bleiben.
+
+**Empfehlung:** Die Error-Message-Logik (`buildErrorMessage`, `buildContentBlockedMessage`, `getBlockReasonExplanations`, `getGenericMediaError`) in eine eigene Klasse auslagern, z.B. `MediaErrorMessageBuilder`.
+
+---
+
+## 5. Keine Tests für `buildErrorMessage()` im Handler
+
+Der neue Test `GoogleProviderBlockedContentTest` testet nur die `checkGeminiFinishReason()`-Methode. Die gesamte Message-Building-Logik im `MediaGenerationHandler` (`buildErrorMessage`, `buildContentBlockedMessage`, `getBlockReasonExplanations`) ist ungetestet.
+
+**Empfehlung:** Unit-Tests für die Error-Message-Logik hinzufügen – besonders für Edge-Cases wie unbekannte Block-Reasons und fehlende Language-Werte.
+
+---
+
+## 6. Reflection-Test für private Methode
+
+Der Test nutzt `ReflectionMethod` um die private `checkGeminiFinishReason()` zu testen. Das ist fragil – Refactoring (Umbenennung, Signatur-Änderung) bricht den Test ohne Compiler-Warnung.
+
+**Empfehlung:** Akzeptabel als Pragmatismus, aber idealerweise über die öffentliche API testen (z.B. `generateImage()` mit gemocktem HTTP-Client).
+
+---
+
+## 7. Deutsche Grammatik: Genitiv-Endung
+
+```php
+$msg = "{$displayName} hat die Erstellung des {$mediaLabel}s mit dem Code **{$reason}** abgelehnt.";
+```
+
+`{$mediaLabel}s` erzeugt „des Bilds" – grammatisch korrekt aber unüblich. Üblicher wäre „des Bildes". Für „Audio" und „Video" funktioniert das „s" korrekt.
+
+---
+
+## Zusammenfassung
+
+| # | Schwere | Problem |
+|---|---------|---------|
+| 1 | **Hoch** | `chat()`/`chatStream()` ignorieren blocked content – stiller Bug |
+| 2 | **Hoch** | `generateVideo()` nutzt das neue Exception-Pattern nicht |
+| 3 | **Mittel** | Hardcoded Translations statt i18n-System |
+| 4 | **Mittel** | Handler überschreitet 500-Zeilen-Grenze |
+| 5 | **Mittel** | Keine Tests für Error-Message-Building |
+| 6 | **Niedrig** | Reflection-Test ist fragil |
+| 7 | **Niedrig** | Deutsche Grammatik (Genitiv) |
